fe/requirements.txt
streamlit==1.39.0
streamlit-authenticator==0.4.1
pandas==2.2.3
google-cloud-storage==2.14.0
loguru==0.7.2
mysql-connector-python==9.1.0
xlsxwriter==3.2.0
python-dotenv==1.0.1
tomli==2.1.0
google-cloud-pubsub==2.18.4
openpyxl